factor some stuff out into SnackManager
