+++ Snackify/Models/User.swift
@@ -8,9 +8,8 @@
 
 import Foundation
 
-class User: Codable {
-    // MARK: - Properties
-    
+struct User: Codable {
+    let id: Int
     let firstName: String
     let lastName: String
     let username: String
@@ -22,17 +21,4 @@ class User: Codable {
     let zipCode: String
     let organization: Organization
     var isAdmin: Bool
-    
-    init(firstName: String, lastName: String, username: String, email: String, phoneNumber: String, streetAddress: String, state: String, zipCode: String, organization: Organization) {
-        self.firstName = firstName
-        self.lastName = lastName
-        self.username = username
-        self.email = email
-        self.phoneNumber = phoneNumber
-        self.streetAddress = streetAddress
-        self.state = state
-        self.zipCode = zipCode
-        self.organization = organization
-        self.isAdmin = false
-    }
 }
